+++ pytorch_lightning/plugins/training_type/sharded.py
@@ -28,15 +28,15 @@ if _FAIRSCALE_AVAILABLE:
 
 class DDPShardedPlugin(DDPPlugin):
 
-    REDUCE_BUFFER_SIZE_DEFAULT = 2**23
+    _REDUCE_BUFFER_SIZE_DEFAULT = 2**23  # 8M
 
     def configure_ddp(self):
         self._wrap_optimizers()
         self._model = ShardedDataParallel(
             LightningShardedDataParallel(self.model),
             sharded_optimizer=self.lightning_module.trainer.optimizers,
-            # For multi-nodes, enabling bucketing would improve performance.
-            reduce_buffer_size=self.REDUCE_BUFFER_SIZE_DEFAULT if self.num_nodes > 1 else 0,
+            # For multi-node training, enabling bucketing will improve performance.
+            reduce_buffer_size=self._REDUCE_BUFFER_SIZE_DEFAULT if self.num_nodes > 1 else 0,
         )
 
     def _reinit_optimizers_with_oss(self):
@@ -46,12 +46,12 @@ class DDPShardedPlugin(DDPPlugin):
                 optimizer = optimizer._optimizer
             if not isinstance(optimizer, OSS):
                 optim_class = type(optimizer)
-                is_fp16 = (self.lightning_module.trainer.accelerator_connector.precision == 16)
+                is_fp16 = self.lightning_module.trainer.precision == 16
                 zero_optimizer = OSS(
                     params=optimizer.param_groups,
                     optim=optim_class,
-                    # For multi-nodes, compress the model shards in fp16 before broadcasting
-                    # would improve performance. When using PyTorch AMP, it would not degrade
+                    # For multi-node training, compressing the model shards in fp16 before broadcasting
+                    # improves performance. When using PyTorch AMP, it will not degrade
                     # the model performance.
                     broadcast_fp16=is_fp16 and self.num_nodes > 1,
                     **optimizer.defaults